Signed-off-by: Ashton Larkin <ashton@openrobotics.org>

<!-- 
Please remove the appropriate section.
For example, if this is a new feature, remove all sections except for the "New feature" section

If this is your first time opening a PR, be sure to check the contribution guide:
https://ignitionrobotics.org/docs/all/contributing
-->

# 🦟 Bug fix

Related to https://github.com/ignitionrobotics/ign-physics/issues/219.

## Summary
This PR is motivated by #656 and https://github.com/ignitionrobotics/ign-physics/issues/219.

https://github.com/ignitionrobotics/ign-physics/issues/219 points out that the `UpdateSim` method in the physics system [goes over all links](https://github.com/ignitionrobotics/ign-gazebo/blob/86e6ad93581362eba4e89a58dc1fef058806a081/src/systems/physics/Physics.cc#L1702) since the physics system doesn't know which links changed. For non-static links, we are [always performing pose updates](https://github.com/ignitionrobotics/ign-gazebo/blob/86e6ad93581362eba4e89a58dc1fef058806a081/src/systems/physics/Physics.cc#L1726-L1774), even if the pose hasn't changed. I'm proposing that we cache each non-static link's pose from the previous iteration and only perform a pose update if the link's pose is different from the previous iteration.

Testing this change by profiling a world as done in #656 (3000 non-static simple shapes, unbounded RTF, headless, and TPE for the physics engine) seems to result in a noticeable performance improvement.

Before link caching, `UpdateSim` takes around `8.25ms`. `Local pose`, which contains the pose updates, takes around `5.2ms`, which is more than half of the `UpdateSim` time:

![before_link_caching](https://user-images.githubusercontent.com/42042756/110020452-8e3c1280-7cf7-11eb-8b9e-c181644bde08.png)

After link caching, `UpdateSim` takes around `5.7ms`, and `Local pose` now takes around `2.4ms`:

![after_link_caching](https://user-images.githubusercontent.com/42042756/110021325-93e62800-7cf8-11eb-8cae-9deb6a9345b7.png)

Taking a look at the some of the other code in `UpdateSim`, I am wondering if we can do something similar for non-static link velocities and accelerations. @chapulina what do you think about this?

## Checklist
- [x] Signed all commits for DCO
- [x] Added tests
- [x] Updated documentation (as needed)
- [x] Updated migration guide (as needed)
- [x] `codecheck` passed (See [contributing](https://ignitionrobotics.org/docs/all/contributing#contributing-code))
- [x] All tests passed (See [test coverage](https://ignitionrobotics.org/docs/all/contributing#test-coverage))
- [x] While waiting for a review on your PR, please help review [another open pull request](https://github.com/pulls?q=is%3Aopen+is%3Apr+user%3Aignitionrobotics+repo%3Aosrf%2Fsdformat+archived%3Afalse+) to support the maintainers

**Note to maintainers**: Remember to use **Squash-Merge**